backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Supabase client (downgraded for compatibility)
supabase==2.0.3

# HTTP client - compatible with supabase-py
httpx>=0.27.0,<0.28

# Required dependencies (supabase-py uses these under the hood)
httpcore>=1.0.0,<2.0
websockets>=12.0,<13.0

# OpenAI for transcription and summarization
openai==1.3.0

# FastAPI dependencies
python-multipart==0.0.6
pydantic==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# Text processing and chunking
langchain==0.1.0
